Spreading things out a bit here, migrated all the database-specific commands from the now-deleted fabfile.py into the new db.py file. (Note: I'm totally not using the @roles decorator properly, I should be using @hosts.) Pulled the rest from fabfile.py into the fabfile/__init__.py file.
As for the rest, added a single deploy/release scheme, but it's pretty darn primitive, in that it only is geared for a very specific set up. Documentation on this type of arrangement is needed now.

Also, changed some of the command names, and need to update the README file to reflect that. Primarly though our new command set revolves around best-practices naming-convention, so tasks are named migrate, update, deploy, and dump.